Removed legacy Hostnode support
In order to align operational model, host nodes are now operated operated outside the labs management. For the initial implementation it'll be sufficient to run a CLI tool to manually setup new VM and re-integrate 10xeng-node cookbook with new nodes API.

Later on, we can use API statistics to setup auto-scaling.